utlimos reportes finalizados

+++ QuizzWebApp-arias/src/main/java/co/uniquindio/edu/quizzwebapp/repositories/AdministradorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdministradorRepository extends JpaRepository<Administrador, Long> {
+public interface AdministradorRepository extends JpaRepository<Administrador, Integer> {
 
 
     Administrador findByCorreoAndPassword(String correo, String password);
