+++ QuizzWebApp-arias/src/main/java/co/uniquindio/edu/quizzwebapp/repositories/DocenteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocenteRepository extends JpaRepository<Docente, Long> {
+public interface DocenteRepository extends JpaRepository<Docente, Integer> {
 
         Docente findByCorreoAndPassword(String correo, String password);
 }
