+++ QuizzWebApp-arias/src/main/java/co/uniquindio/edu/quizzwebapp/repositories/EstudianteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EstudianteRepository extends JpaRepository<Estudiante, Long> {
+public interface EstudianteRepository extends JpaRepository<Estudiante, Integer> {
 
     Estudiante findByCorreoAndPassword(String correo, String password);
 }
